added a count of total bookmarks on each page
The patch adds a count of the total number of bookmarks that would be
displayed if there were not a per-page limit for the current page. It is
added in parentheses to the subheader on each page or not shown at all
if the variable is not defined for the page.