Added ability to set default ordering of bookmarks to modified date
This allows you to set default bookmark ordering to be last modified
date, so if you bookmark a URL and it is already bookmarked and you
save it again, then it will appear with most recent bookmarks, instead
of three years ago, for instance. Set this option in config.inc.php.